Rename module + copyright notice cleanup
